Refactor chart range extensions and controls
Moved range-related files from Event/Range to Chart/Range for better organization. Split ChartExtent+Enabled logic into a new file and removed duplicate code from ChartExtent+Move.
