+++ Sources/Scout/UI/Chart/Range/Range+Label.swift
